requirements.txt
agents==1.4.0
openai==1.68.2
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.2